Move hooks to separate folder
