+++ src/components/WalletModal/LenderVaults/LenderVaults.tsx
@@ -1,8 +1,5 @@
-import { useEffect, useState } from 'react'
-
 import { useConnection, useWallet } from '@solana/wallet-adapter-react'
 import classNames from 'classnames'
-import { BN } from 'fbonds-core'
 import { LendingTokenType } from 'fbonds-core/lib/fbond-protocol/types'
 import { uniqueId } from 'lodash'
 import { TxnExecutor } from 'solana-transactions-executor'
@@ -13,7 +10,6 @@ import { DisplayValue } from '@banx/components/TableComponents'
 import Tooltip from '@banx/components/Tooltip'
 import { InputErrorMessage, NumericStepInput } from '@banx/components/inputs'
 
-import { useWalletBalance } from '@banx/hooks'
 import { useTokenType } from '@banx/store/common'
 import {
   TXN_EXECUTOR_DEFAULT_OPTIONS,
@@ -22,139 +18,37 @@ import {
 } from '@banx/transactions'
 import {
   CreateClaimLenderVaultTxnDataParams,
-  CreateUpdateUserVaultTxnDataParams,
   createClaimLenderVaultTxnData,
-  createUpdateUserVaultTxnData,
   parseClaimLenderVaultSimulatedAccounts,
-  parseDepositSimulatedAccounts,
 } from '@banx/transactions/vault'
 import {
-  ZERO_BN,
   destroySnackbar,
   enqueueConfirmationError,
   enqueueSnackbar,
   enqueueTransactionsSent,
   enqueueWaitingConfirmation,
-  getTokenDecimals,
   getTokenUnit,
   isBanxSolTokenType,
-  stringToBN,
 } from '@banx/utils'
 
 import { TooltipRow } from '../components'
 import { BanxSolEpochContent, EscrowTabs } from './components'
-import { getInputErrorMessage } from './helpers'
-import { useLenderVaultInfo } from './hooks'
+import { useLenderVaultInfo, useUserVaultContent } from './hooks'
 
 import styles from './LenderVaults.module.less'
 
 export const EscrowVault = () => {
-  const wallet = useWallet()
-  const { connection } = useConnection()
-  const { tokenType } = useTokenType()
-
-  const { lenderVaultInfo, updateUserVaultOptimistic } = useLenderVaultInfo()
-  const walletBalance = useWalletBalance(tokenType)
-
-  const [activeTab, setActiveTab] = useState<'wallet' | 'escrow'>('wallet')
-  const [inputValue, setInputValue] = useState('0')
-
-  const tokenDecimals = getTokenDecimals(tokenType)
-
-  const formatBalance = (balance: number) => (balance / tokenDecimals).toString()
-
-  useEffect(() => {
-    setInputValue(formatBalance(walletBalance))
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const onTabClick = (nextTab: 'wallet' | 'escrow') => {
-    setActiveTab(nextTab)
-
-    const balance = nextTab === 'wallet' ? walletBalance : lenderVaultInfo.offerLiquidityAmount
-    setInputValue(formatBalance(balance))
-  }
-
-  const update = async (amount: BN) => {
-    if (amount.lt(ZERO_BN)) return
-
-    const loadingSnackbarId = uniqueId()
-
-    try {
-      const walletAndConnection = createExecutorWalletAndConnection({ wallet, connection })
-
-      const txnData = await createUpdateUserVaultTxnData(
-        {
-          amount,
-          lendingTokenType: tokenType,
-          add: activeTab === 'wallet',
-        },
-        walletAndConnection,
-      )
-
-      await new TxnExecutor<CreateUpdateUserVaultTxnDataParams>(walletAndConnection, {
-        ...TXN_EXECUTOR_DEFAULT_OPTIONS,
-      })
-        .addTxnData(txnData)
-        .on('sentAll', () => {
-          enqueueTransactionsSent()
-          enqueueWaitingConfirmation(loadingSnackbarId)
-        })
-        .on('confirmedAll', (results) => {
-          const { confirmed, failed } = results
-
-          destroySnackbar(loadingSnackbarId)
-
-          if (confirmed.length) {
-            enqueueSnackbar({
-              message: activeTab === 'wallet' ? 'Successfully deposited' : 'Successfully withdrawn',
-              type: 'success',
-            })
-            confirmed.forEach(({ accountInfoByPubkey }) => {
-              if (!accountInfoByPubkey) return
-              const userVault = parseDepositSimulatedAccounts(accountInfoByPubkey)
-
-              updateUserVaultOptimistic({
-                walletPubkey: walletAndConnection.wallet.publicKey.toBase58(),
-                updatedUserVault: userVault,
-              })
-            })
-          }
-
-          if (failed.length) {
-            return failed.forEach(({ signature, reason }) =>
-              enqueueConfirmationError(signature, reason),
-            )
-          }
-        })
-        .on('error', (error) => {
-          throw error
-        })
-        .execute()
-    } catch (error) {
-      destroySnackbar(loadingSnackbarId)
-      defaultTxnErrorHandler(error, {
-        additionalData: {
-          amount: inputValue,
-          lendingTokenType: tokenType,
-        },
-        walletPubkey: wallet?.publicKey?.toBase58(),
-        transactionName: 'UpdateUserVault',
-      })
-    }
-  }
-
-  const onActionClick = () => {
-    update(stringToBN(inputValue, Math.log10(tokenDecimals)))
-  }
-
-  const errorMessage = getInputErrorMessage({
-    activeTab,
-    walletBalance,
-    escrowBalance: lenderVaultInfo.offerLiquidityAmount,
+  const {
     inputValue,
+    setInputValue,
+    activeTab,
+    onTabClick,
+    onActionClick,
+    walletBalance,
+    escrowBalance,
+    errorMessage,
     tokenType,
-  })
+  } = useUserVaultContent()
 
   return (
     <div className={styles.wrapper}>
@@ -166,7 +60,7 @@ export const EscrowVault = () => {
         tab={activeTab}
         setTab={onTabClick}
         walletBalance={walletBalance}
-        escrowBalance={lenderVaultInfo.offerLiquidityAmount}
+        escrowBalance={escrowBalance}
       />
 
       <NumericStepInput
