+++ src/components/WalletModal/LenderVaults/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useUserVault'
+export * from './useUserVaultContent'
